Move button component into inputs folder
