+++ src/components/Input/index.js
@@ -1,5 +1,6 @@
+import Button from './Button';
 import Label from './Label';
 import Location from './Location';
 import Select from './Select';
 
-export { Label, Location, Select };
+export { Button, Label, Location, Select };
